+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { Route, Routes } from "react-router-dom";
-import CategoryDetailPage from "./components/CategoryDetailPage";
-import CategoryPage from "./components/CategoryPage";
-import PageLayout from "./components/PageLayout";
-import StoreDetailPage from "./components/StoreDetailPage";
+import CategoryDetailPage from "./components/pages/CategoryDetailPage";
+import CategoryPage from "./components/pages/CategoryPage";
+import PageLayout from "./components/layout/PageLayout";
+import StoreDetailPage from "./components/pages/StoreDetailPage";
 
 function App() {
   return (
