+++ src/components/common/Star.tsx
