+++ src/components/common/StoreListItem.tsx
