+++ src/stories/Header.stories.jsx
@@ -1,4 +1,4 @@
-import Header from "../components/Header";
+import Header from "../components/layout/Header";
 
 export default {
   title: "Component/Header",
