+++ src/stories/SectionHeader.stories.jsx
@@ -1,6 +1,6 @@
 import { MdArrowBackIos } from "react-icons/md";
 
-import SectionHeader from "../components/SectionHeader";
+import SectionHeader from "../components/common/SectionHeader";
 
 export default {
   title: "Component/SectionHeader",
